+++ NEWS.md
@@ -7,6 +7,9 @@
 creates/deletes/renames/pastes a node.
 * Added new Shiny value `outputId_move` triggered when a node is moved.
 * Added new Shiny value `outputId_rename` triggered when a node is renamed.
+* Added new Shiny value `outputId_paste` triggered when a node is copied and 
+pasted.
+
 
 
 # jsTreeR 0.1.0 (2020-08-26)